refactor convention
