+++ .vscode/settings.json
@@ -23,5 +23,6 @@
   },
   "javascript.preferences.importModuleSpecifier": "relative",
   "typescript.preferences.importModuleSpecifier": "relative",
+  "editor.formatOnSave": true,
   "prettier.configPath": ".prettierrc.js",
 }
